\name{print.coxinterval}
\alias{print.coxinterval}

\title{Print method for model fit}

\description{
  Prints an object the class \code{"coxinterval"}.
}

\usage{
\method{print}{coxinterval}(x, ...)
}

\arguments{
  \item{x}{
    an object returned by \code{\link{coxaalen}} or \code{\link{coxdual}}.
  }
  \item{\dots}{
    further arguments for other methods.
  }
}

\seealso{
  \code{\link{summary.coxinterval}}
}

\keyword{survival}
